Set $TERM, $CLICOLOR, $SUBLIME vars in subshell
If your .bash_profile script does special stuff that may not be friendly
to executing in a subshell under Sublime (for example, echoing ANSI
control characters or the like,) adding a `if [[ $SUBLIME == 1 ]];
then... fi` block around those commands should help alleviate that.

We also set $TERM to 'ansi' so that we get the most basic, ANSI-standard
terminal output we can; and $CLICOLOR to an empty string ("") to turn
off ANSI color codes in the subshell.