Strip ANSI control chars from output. Fixes #3.
This plugin runs a login subshell and echoes $PATH from within it to
find the natual value of $PATH. However, as described in #3, some users
automatically print characters which mess with our parsing of the $PATH,
namely ANSI control characters.

This commit explicitly strips ANSI control characters from the output of
our `printf "$PATH"` command, using a regular expression in Python. The
regular expression in question was taken from
http://www.commandlinefu.com/commands/view/3584/remove-color-codes-special-characters-with-sed
.