Don't run if OS isn't Mac OS X (fixes #5)
We add an explicit check before loading the plugin to make sure the
current operating system is Mac OS X. If this check fails, none of the
plugin code is run (and thus, the path isn't touched) and a message is
printed to the Sublime console (View -> Show Console).

This will prevent this plugin from corrupting the path on Windows and
Linux as described in #5. As should be obvious from the plugin name and
README, this plugin is only designed to work on Mac OS X.