Fix Python 2/Sublime Text 2 support
7348deb07092e25b prints out a debugging message if the OS isn't OS X.
However, we used the Python 3-style `print()` command to do so. Since
Sublime Text 2 uses Python 2, we add a bevy of `from __future__
import...` at the top of the Python file, so that we get Python
3 features in Python 2.

This should hopefully fix Python 2/Sublime Text 2 problems, but it is
untested. Feedback appreciated.